pyproject.toml
[tool.poetry]
name = "data-viewer"
version = "0.1.0"
description = "View files in human-readable format."
readme = "README.md"
authors = ["lamcw <thomas@lamcw.com>"]

[tool.poetry.scripts]
data-viewer = "data_viewer.console:main"

[tool.poetry.dependencies]
python = "^3.8"
dacite = "^1.4.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"
pydocstyle = "^5.0.2"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
